Hotkeys become key combinations; default to Ctrl+Win
Three things, all from one report: Wispr Flow uses Ctrl+Win, multiple keys
should be bindable, and holding Shift raises the Windows Filter Keys prompt.

The Filter Keys point is a defect I introduced. Right Shift was offered as a
push-to-talk key, and holding either Shift for eight seconds opens that prompt -
a push-to-talk hold routinely runs longer than that. No Shift-only preset is
offered now. Shift can still be part of a recorded combination, with a warning:
it is the user's keyboard, and someone who has already turned Filter Keys off
should not be blocked.

PushToTalkKey (one key) becomes Hotkey (a set held together), with a picker of
presets and a Record button that captures whatever the user holds. Capture is
reported on release of the last key rather than the first press, and takes the
largest set held at once - people do not press every key of a chord on the same
millisecond, and reading the final state would capture whatever happened to
still be down.

Ctrl+Win is the default and the chord is load-bearing: Win alone opens the Start
menu on release, and Ctrl held with it makes Windows treat the pair as a chord
so it does not. Verified against the real hook - foreground window unchanged
across a full press and release.

All the combination logic is in Teezy.Core behind HotkeyMatcher, because that is
where the bugs are: press order, auto-repeat, partial release, and a key held on
both sides at once. Each slot tracks WHICH physical keys satisfy it, not merely
whether one does - hold both Ctrl keys, release one, and a Ctrl is still down,
so the hold must not end mid-sentence. The Win32 layer only translates key
events into HotkeyKey values.

Old settings files migrate rather than reset: silently changing someone's hotkey
would be worse than any amount of migration code. RightShift migrates faithfully
to a Shift combination even though it is now discouraged - a migration that
changes behaviour is a surprise, not a migration. The legacy field is dropped on
write, so the file converts itself on the first save.

105 tests, including thirteen on the matcher alone and eight on migration.
Verified against the real hook: chord fires on the second key, neither key alone
fires, either side satisfies a side-agnostic slot, an unrelated key tapped
mid-hold does not break it, and auto-repeat yields exactly one press.

+++ src/Teezy.Platform.Windows/WindowsTextInjector.cs
@@ -102,5 +102,5 @@ public sealed class WindowsTextInjector : ITextInjector
 
     /// <summary>The high bit of <c>GetAsyncKeyState</c> is the physical down state; the low
     /// bit is "pressed since last call" and must not be tested here.</summary>
-    private static bool IsDown(int vk) => (GetAsyncKeyState(vk) & 0x8000) != 0;
+    private static bool IsDown(uint vk) => (GetAsyncKeyState((int)vk) & 0x8000) != 0;
 }
